+++ unpacker.go
@@ -145,18 +145,15 @@ func (upckr *unpacker) unpackObject() (interface{}, error) {
 	case 0xc2:
 		return false, nil
 
-		// TODO: Float support for Value?
-	// case 0xca: {
-	//   val = Float.intBitsToFloat(Buffer.bytesToInt(upckr.buffer, upckr.offset));
-	//   upckr.offset += 4;
-	//   return upckr.GetDouble(val);
-	// }
+	case 0xca:
+		val := Buffer.BytesToFloat32(upckr.buffer, upckr.offset)
+		upckr.offset += 4
+		return val, nil
 
-	// case 0xcb: {
-	//   double val = Double.longBitsToDouble(Buffer.bytesToLong(upckr.buffer, upckr.offset));
-	//   upckr.offset += 8;
-	//   return upckr.GetDouble(val);
-	// }
+	case 0xcb:
+		val := Buffer.BytesToFloat64(upckr.buffer, upckr.offset)
+		upckr.offset += 8
+		return val, nil
 
 	case 0xcc:
 		r := upckr.buffer[upckr.offset] & 0xff
@@ -178,11 +175,10 @@ func (upckr *unpacker) unpackObject() (interface{}, error) {
 		}
 		return int64(val), nil
 
-	// TODO: Fix upckr
 	case 0xcf:
 		val := Buffer.BytesToInt64(upckr.buffer, upckr.offset)
 		upckr.offset += 8
-		return val, nil
+		return uint64(val), nil
 
 	case 0xd0:
 		r := int8(upckr.buffer[upckr.offset])
@@ -254,7 +250,6 @@ func (upckr *unpacker) unpackObject() (interface{}, error) {
 		if theType >= 0xe0 {
 			return int(theType - 0xe0 - 32), nil
 		}
-		// panic(fmt.Errorf("Unknown upckr.unpack theType: %x", theType))
 	}
 
 	return nil, NewAerospikeError(SERIALIZE_ERROR)